+++ pages/dao/[id]/activity.tsx
@@ -18,6 +18,10 @@ import StarIcon from "@mui/icons-material/Star";
 import Activity, { IActivity } from "@components/dao/activity/Activity";
 import Musk from "@public/profile/musk-full.png";
 import PaideiaLogo from "@public/dao/bio-image/paideia-logo.png";
+import { paths, props } from "@lib/DaoPaths";
+
+export const getStaticPaths = paths;
+export const getStaticProps = props;
 
 const categories = [
   { icon: <AppsIcon sx={{ mr: ".2rem", fontSize: ".9rem" }} />, label: "All" },
