+++ pages/dao/[id]/member/[member_id].tsx
@@ -1,9 +1,9 @@
-import { Box } from "@mui/material";
 import * as React from "react";
 import { paths, props } from "@lib/MemberPaths";
+import AbstractProfile from "@components/dao/profile/AbstractProfile";
 
 const Member: React.FC = () => {
-  return <Box sx={{ p: "1.5rem" }}>Member Here...</Box>;
+  return <AbstractProfile/>;
 };
 export default Member;
 export const getStaticPaths = paths;
