+++ pages/dao/[id]/wallet.tsx
@@ -15,8 +15,12 @@ import dateFormat from "dateformat";
 import { CheckCircle } from "@mui/icons-material";
 import PaideiaLogo from "@public/dao/bio-image/paideia-logo.png";
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-
 import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
+import { paths, props } from "@lib/DaoPaths";
+
+export const getStaticPaths = paths;
+export const getStaticProps = props;
+
 
 const ActiveWallet: React.FC<{ previous?: boolean }> = (props) => {
   const { wallet, dAppWallet } = useWallet();
